requirements.txt
Django==3.2.7
psycopg2-binary>=2.8
djangorestframework==3.12.4
django-rest-swagger==2.1.2
djoser==2.1.0